backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
python-dotenv>=1.0
sse-starlette>=2.0
openai>=1.30.0
httpx>=0.27.0
Pillow>=10.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-pptx>=0.6.23
pytesseract>=0.3.10
